remaining typescript, component rename, reorg
