+++ src/components/EmojiPaint/Footer.tsx
@@ -1,0 +1,32 @@
+import React from 'react';
+import '../../index.css';
+
+interface ToolBarProps {
+  clearEmoji: () => void;
+  disabled: boolean;
+  copyMessage: () => void;
+}
+export const Footer = ({
+  clearEmoji,
+  disabled,
+  copyMessage,
+}: ToolBarProps): JSX.Element => {
+  return (
+    <div className="emoji-grid__toolbar">
+      <div className="emoji-grid__controls">
+        <button
+          className="emoji-grid__control"
+          onClick={clearEmoji}
+          disabled={disabled}>
+          <span>Clear</span>
+        </button>
+        <button
+          className="emoji-grid__control btn_green"
+          onClick={copyMessage}
+          disabled={disabled}>
+          <span>Copy to Clipboard</span>
+        </button>
+      </div>
+    </div>
+  );
+};
